requirements.txt
aiohttp
yandex-music
aiogram
psycopg2-binary